Refactored client side.

+++ HomeworkAssignment.Persistence/Commands/Assignments/UpdatePartialAssignmentCommand.cs
@@ -3,4 +3,4 @@ using MediatR;
 
 namespace HomeAssignment.Persistence.Commands.Assignments;
 
-public sealed record UpdatePartialAssignmentCommand(Guid Id, int Position) : IRequest;
+public sealed record UpdatePartialAssignmentCommand(Guid Id, ushort Position) : IRequest;
