+++ HomeworkAssignment.Persistence/Commands/Chapters/UpdatePartialChapterCommand.cs
@@ -2,4 +2,4 @@
 
 namespace HomeAssignment.Persistence.Commands.Chapters;
 
-public sealed record UpdatePartialChapterCommand(Guid ChapterId, int Position) : IRequest;
+public sealed record UpdatePartialChapterCommand(Guid ChapterId, ushort Position) : IRequest;
